split main.js on components
