+++ src/components/filmCards.js
@@ -1,0 +1,9 @@
+import {renderCard} from '@components/card.js';
+
+const renderFilmCards = function (container, data) {
+  data.forEach((it) => {
+    renderCard(container, it);
+  });
+};
+
+export {renderFilmCards};
