+++ src/components/rank.js
@@ -1,0 +1,15 @@
+import {render} from '@components/utils.js';
+
+const renderRank = function (container) {
+  const createTemplate = () => {
+    return (
+      `<section class="header__profile profile">
+      <p class="profile__rating">Movie Buff</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>`
+    );
+  };
+  render(container, createTemplate(), `beforeEnd`);
+};
+
+export {renderRank};
