+++ src/components/utils.js
@@ -1,0 +1,3 @@
+export const render = (container, template, targetPlace) => {
+  container.insertAdjacentHTML(targetPlace, template);
+};
